Create setUserData action. Refactor getUserData. Fill out and export UserProvider component
